requirements.txt
bs4
tqdm
emoji
flask
jinja2
gputil
pillow
psutil
requests
ansicolors
cefpython3
python-dotenv